+++ src/apps/testapps/testH3SetToLinkedGeo.c
@@ -17,35 +17,24 @@
 #include <stdio.h>
 #include "algos.h"
 #include "test.h"
-
-H3Index* makeSet(char** hexes, int numHexes) {
-    H3Index* set = calloc(numHexes, sizeof(H3Index));
-    for (int i = 0; i < numHexes; i++) {
-        set[i] = H3_EXPORT(stringToH3)(hexes[i]);
-    }
-    return set;
-}
+#include "utility.h"
 
 BEGIN_TESTS(h3SetToLinkedGeo);
 
 TEST(empty) {
     LinkedGeoPolygon polygon;
-    int numHexes = 0;
-    H3Index* set = makeSet(NULL, numHexes);
 
-    H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
+    H3_EXPORT(h3SetToLinkedGeo)(NULL, 0, &polygon);
 
     t_assert(countLinkedLoops(&polygon) == 0, "No loops added to polygon");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(singleHex) {
     LinkedGeoPolygon polygon;
-    char* hexes[] = {"890dab6220bffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x890dab6220bffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
@@ -53,14 +42,12 @@ TEST(singleHex) {
     t_assert(countLinkedCoords(polygon.first) == 6, "6 coords added to loop");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(contiguous2) {
     LinkedGeoPolygon polygon;
-    char* hexes[] = {"8928308291bffff", "89283082957ffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x8928308291bffff, 0x89283082957ffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
@@ -69,7 +56,6 @@ TEST(contiguous2) {
              "All coords added to loop except 2 shared");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 // TODO: This test asserts incorrect behavior - we should be creating multiple
@@ -77,27 +63,27 @@ TEST(contiguous2) {
 // corrected.
 TEST(nonContiguous2) {
     LinkedGeoPolygon polygon;
-    char* hexes[] = {"8928308291bffff", "89283082943ffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x8928308291bffff, 0x89283082943ffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
-    t_assert(countLinkedLoops(&polygon) == 2, "2 loops added to polygon");
+    t_assert(countLinkedPolygons(&polygon) == 2, "2 polygons added");
+    t_assert(countLinkedLoops(&polygon) == 1, "1 loop on the first polygon");
     t_assert(countLinkedCoords(polygon.first) == 6,
              "All coords for one hex added to first loop");
-    t_assert(countLinkedCoords(polygon.first->next) == 6,
-             "All coords for one hex added to second loop");
+    t_assert(countLinkedLoops(polygon.next) == 1,
+             "Loop count on second polygon correct");
+    t_assert(countLinkedCoords(polygon.next->first) == 6,
+             "All coords for one hex added to second polygon");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(contiguous3) {
     LinkedGeoPolygon polygon;
-    char* hexes[] = {"8928308288bffff", "892830828d7ffff", "8928308289bffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x8928308288bffff, 0x892830828d7ffff, 0x8928308289bffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
@@ -106,35 +92,29 @@ TEST(contiguous3) {
              "All coords added to loop except 6 shared");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(hole) {
     LinkedGeoPolygon polygon;
-    char* hexes[] = {"892830828c7ffff", "892830828d7ffff", "8928308289bffff",
-                     "89283082813ffff", "8928308288fffff", "89283082883ffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x892830828c7ffff, 0x892830828d7ffff, 0x8928308289bffff,
+                     0x89283082813ffff, 0x8928308288fffff, 0x89283082883ffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
     t_assert(countLinkedLoops(&polygon) == 2, "2 loops added to polygon");
-    // Note: This isn't strictly correct, and should be reversed when
-    // https://github.com/uber/h3/issues/53 is resolved
-    t_assert(countLinkedCoords(polygon.first) == 6,
-             "All inner coords added to first loop");
-    t_assert(countLinkedCoords(polygon.first->next) == 6 * 3,
-             "All outer coords added to second loop");
+    t_assert(countLinkedCoords(polygon.first) == 6 * 3,
+             "All outer coords added to first loop");
+    t_assert(countLinkedCoords(polygon.first->next) == 6,
+             "All inner coords added to second loop");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(pentagon) {
     LinkedGeoPolygon polygon;
-    char* hexes[] = {"851c0003fffffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x851c0003fffffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
@@ -143,21 +123,19 @@ TEST(pentagon) {
              "10 coords (distorted pentagon) added to loop");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(2Ring) {
     LinkedGeoPolygon polygon;
     // 2-ring, in order returned by k-ring algo
-    char* hexes[] = {"8930062838bffff", "8930062838fffff", "89300628383ffff",
-                     "8930062839bffff", "893006283d7ffff", "893006283c7ffff",
-                     "89300628313ffff", "89300628317ffff", "893006283bbffff",
-                     "89300628387ffff", "89300628397ffff", "89300628393ffff",
-                     "89300628067ffff", "8930062806fffff", "893006283d3ffff",
-                     "893006283c3ffff", "893006283cfffff", "8930062831bffff",
-                     "89300628303ffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x8930062838bffff, 0x8930062838fffff, 0x89300628383ffff,
+                     0x8930062839bffff, 0x893006283d7ffff, 0x893006283c7ffff,
+                     0x89300628313ffff, 0x89300628317ffff, 0x893006283bbffff,
+                     0x89300628387ffff, 0x89300628397ffff, 0x89300628393ffff,
+                     0x89300628067ffff, 0x8930062806fffff, 0x893006283d3ffff,
+                     0x893006283c3ffff, 0x893006283cfffff, 0x8930062831bffff,
+                     0x89300628303ffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
@@ -166,21 +144,19 @@ TEST(2Ring) {
              "Expected number of coords added to loop");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(2RingUnordered) {
     LinkedGeoPolygon polygon;
     // 2-ring in random order
-    char* hexes[] = {"89300628393ffff", "89300628383ffff", "89300628397ffff",
-                     "89300628067ffff", "89300628387ffff", "893006283bbffff",
-                     "89300628313ffff", "893006283cfffff", "89300628303ffff",
-                     "89300628317ffff", "8930062839bffff", "8930062838bffff",
-                     "8930062806fffff", "8930062838fffff", "893006283d3ffff",
-                     "893006283c3ffff", "8930062831bffff", "893006283d7ffff",
-                     "893006283c7ffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x89300628393ffff, 0x89300628383ffff, 0x89300628397ffff,
+                     0x89300628067ffff, 0x89300628387ffff, 0x893006283bbffff,
+                     0x89300628313ffff, 0x893006283cfffff, 0x89300628303ffff,
+                     0x89300628317ffff, 0x8930062839bffff, 0x8930062838bffff,
+                     0x8930062806fffff, 0x8930062838fffff, 0x893006283d3ffff,
+                     0x893006283c3ffff, 0x8930062831bffff, 0x893006283d7ffff,
+                     0x893006283c7ffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
@@ -189,14 +165,79 @@ TEST(2RingUnordered) {
              "Expected number of coords added to loop");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
+}
+
+TEST(nestedDonut) {
+    LinkedGeoPolygon polygon;
+    // hollow 1-ring + hollow 3-ring around the same hex
+    H3Index set[] = {0x89283082813ffff, 0x8928308281bffff, 0x8928308280bffff,
+                     0x8928308280fffff, 0x89283082807ffff, 0x89283082817ffff,
+                     0x8928308289bffff, 0x892830828d7ffff, 0x892830828c3ffff,
+                     0x892830828cbffff, 0x89283082853ffff, 0x89283082843ffff,
+                     0x8928308284fffff, 0x8928308287bffff, 0x89283082863ffff,
+                     0x89283082867ffff, 0x8928308282bffff, 0x89283082823ffff,
+                     0x89283082837ffff, 0x892830828afffff, 0x892830828a3ffff,
+                     0x892830828b3ffff, 0x89283082887ffff, 0x89283082883ffff};
+    int numHexes = ARRAY_SIZE(set);
+
+    H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
+
+    // Note that the polygon order here is arbitrary, making this test
+    // somewhat brittle, but it's difficult to assert correctness otherwise
+    t_assert(countLinkedPolygons(&polygon) == 2, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 2,
+             "Loop count on first polygon correct");
+    t_assert(countLinkedCoords(polygon.first) == 42,
+             "Got expected big outer loop");
+    t_assert(countLinkedCoords(polygon.first->next) == 30,
+             "Got expected big inner loop");
+    t_assert(countLinkedLoops(polygon.next) == 2,
+             "Loop count on second polygon correct");
+    t_assert(countLinkedCoords(polygon.next->first) == 18,
+             "Got expected outer loop");
+    t_assert(countLinkedCoords(polygon.next->first->next) == 6,
+             "Got expected inner loop");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(nestedDonutTransmeridian) {
+    LinkedGeoPolygon polygon;
+    // hollow 1-ring + hollow 3-ring around the hex at (0, -180)
+    H3Index set[] = {0x897eb5722c7ffff, 0x897eb5722cfffff, 0x897eb572257ffff,
+                     0x897eb57220bffff, 0x897eb572203ffff, 0x897eb572213ffff,
+                     0x897eb57266fffff, 0x897eb5722d3ffff, 0x897eb5722dbffff,
+                     0x897eb573537ffff, 0x897eb573527ffff, 0x897eb57225bffff,
+                     0x897eb57224bffff, 0x897eb57224fffff, 0x897eb57227bffff,
+                     0x897eb572263ffff, 0x897eb572277ffff, 0x897eb57223bffff,
+                     0x897eb572233ffff, 0x897eb5722abffff, 0x897eb5722bbffff,
+                     0x897eb572287ffff, 0x897eb572283ffff, 0x897eb57229bffff};
+    int numHexes = ARRAY_SIZE(set);
+
+    H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
+
+    // Note that the polygon order here is arbitrary, making this test
+    // somewhat brittle, but it's difficult to assert correctness otherwise
+    t_assert(countLinkedPolygons(&polygon) == 2, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 2,
+             "Loop count on first polygon correct");
+    t_assert(countLinkedCoords(polygon.first) == 18, "Got expected outer loop");
+    t_assert(countLinkedCoords(polygon.first->next) == 6,
+             "Got expected inner loop");
+    t_assert(countLinkedLoops(polygon.next) == 2,
+             "Loop count on second polygon correct");
+    t_assert(countLinkedCoords(polygon.next->first) == 42,
+             "Got expected big outer loop");
+    t_assert(countLinkedCoords(polygon.next->first->next) == 30,
+             "Got expected big inner loop");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
 }
 
 TEST(contiguous2distorted) {
     LinkedGeoPolygon polygon;
-    char* hexes[] = {"894cc5365afffff", "894cc536537ffff"};
-    int numHexes = sizeof(hexes) / sizeof(hexes[0]);
-    H3Index* set = makeSet(hexes, numHexes);
+    H3Index set[] = {0x894cc5365afffff, 0x894cc536537ffff};
+    int numHexes = ARRAY_SIZE(set);
 
     H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
 
@@ -205,14 +246,23 @@ TEST(contiguous2distorted) {
              "All coords added to loop except 2 shared");
 
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
-    free(set);
 }
 
 TEST(negativeHashedCoordinates) {
     LinkedGeoPolygon polygon;
-    H3Index set[] = {0x88ad36c547fffffl, 0x88ad36c467fffffl};
-    H3_EXPORT(h3SetToLinkedGeo)(set, 2, &polygon);
-    t_assert(countLinkedLoops(&polygon) == 2, "2 loops added to polygon");
+    H3Index set[] = {0x88ad36c547fffff, 0x88ad36c467fffff};
+    int numHexes = ARRAY_SIZE(set);
+    H3_EXPORT(h3SetToLinkedGeo)(set, numHexes, &polygon);
+
+    t_assert(countLinkedPolygons(&polygon) == 2, "2 polygons added");
+    t_assert(countLinkedLoops(&polygon) == 1, "1 loop on the first polygon");
+    t_assert(countLinkedCoords(polygon.first) == 6,
+             "All coords for one hex added to first loop");
+    t_assert(countLinkedLoops(polygon.next) == 1,
+             "Loop count on second polygon correct");
+    t_assert(countLinkedCoords(polygon.next->first) == 6,
+             "All coords for one hex added to second polygon");
+
     H3_EXPORT(destroyLinkedPolygon)(&polygon);
 }