+++ src/apps/testapps/testPolygon.c
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+#include <assert.h>
 #include <stdlib.h>
 #include "bbox.h"
 #include "constants.h"
@@ -28,61 +29,43 @@ GeoCoord sfVerts[] = {
     {0.659966917655, -2.1364398519396},  {0.6595011102219, -2.1359434279405},
     {0.6583348114025, -2.1354884206045}, {0.6581220034068, -2.1382437718946},
     {0.6594479998527, -2.1384597563896}, {0.6599990002976, -2.1376771158464}};
-Geofence sfGeofence;
 
-GeoCoord primeMeridianVerts[] = {
-    {0.01, 0.01}, {0.01, -0.01}, {-0.01, -0.01}, {-0.01, 0.01}};
-Geofence primeMeridianGeofence;
-
-GeoCoord transMeridianVerts[] = {{0.01, -M_PI + 0.01},
-                                 {0.01, M_PI - 0.01},
-                                 {-0.01, M_PI - 0.01},
-                                 {-0.01, -M_PI + 0.01}};
-Geofence transMeridianGeofence;
-
-GeoCoord transMeridianHoleVerts[] = {{0.005, -M_PI + 0.005},
-                                     {0.005, M_PI - 0.005},
-                                     {-0.005, M_PI - 0.005},
-                                     {-0.005, -M_PI + 0.005}};
-Geofence transMeridianHoleGeofence;
-
-static void destroyLinkedGeoLoop(LinkedGeoLoop* loop) {
-    for (LinkedGeoCoord *currentCoord = loop->first, *nextCoord;
-         currentCoord != NULL; currentCoord = nextCoord) {
-        nextCoord = currentCoord->next;
-        free(currentCoord);
+static void createLinkedLoop(LinkedGeoLoop* loop, GeoCoord* verts,
+                             int numVerts) {
+    initLinkedLoop(loop);
+    for (int i = 0; i < numVerts; i++) {
+        addLinkedCoord(loop, verts++);
     }
 }
 
 BEGIN_TESTS(polygon);
 
-sfGeofence.numVerts = 6;
-sfGeofence.verts = sfVerts;
-
-primeMeridianGeofence.numVerts = 4;
-primeMeridianGeofence.verts = primeMeridianVerts;
-
-transMeridianGeofence.numVerts = 4;
-transMeridianGeofence.verts = transMeridianVerts;
-
-transMeridianHoleGeofence.numVerts = 4;
-transMeridianHoleGeofence.verts = transMeridianHoleVerts;
-
 TEST(pointInsideGeofence) {
+    Geofence geofence = {.numVerts = 6, .verts = sfVerts};
+
+    GeoCoord inside = {0.659, -2.136};
     GeoCoord somewhere = {1, 2};
 
     BBox bbox;
-    bboxFromGeofence(&sfGeofence, &bbox);
+    bboxFromGeofence(&geofence, &bbox);
 
-    t_assert(!pointInsideGeofence(&sfGeofence, &bbox, &sfVerts[0]),
+    t_assert(!pointInsideGeofence(&geofence, &bbox, &sfVerts[0]),
              "contains exact");
-    t_assert(pointInsideGeofence(&sfGeofence, &bbox, &sfVerts[4]),
-             "contains exact4");
-    t_assert(!pointInsideGeofence(&sfGeofence, &bbox, &somewhere),
+    t_assert(pointInsideGeofence(&geofence, &bbox, &sfVerts[4]),
+             "contains exact 4");
+    t_assert(pointInsideGeofence(&geofence, &bbox, &inside),
+             "contains point inside");
+    t_assert(!pointInsideGeofence(&geofence, &bbox, &somewhere),
              "contains somewhere else");
 }
 
 TEST(pointInsideGeofenceTransmeridian) {
+    GeoCoord verts[] = {{0.01, -M_PI + 0.01},
+                        {0.01, M_PI - 0.01},
+                        {-0.01, M_PI - 0.01},
+                        {-0.01, -M_PI + 0.01}};
+    Geofence transMeridianGeofence = {.numVerts = 4, .verts = verts};
+
     GeoCoord eastPoint = {0.001, -M_PI + 0.001};
     GeoCoord eastPointOutside = {0.001, -M_PI + 0.1};
     GeoCoord westPoint = {0.001, M_PI - 0.001};
@@ -108,11 +91,7 @@ TEST(pointInsideLinkedGeoLoop) {
     GeoCoord inside = {0.659, -2.136};
 
     LinkedGeoLoop loop;
-    initLinkedLoop(&loop);
-
-    for (int i = 0; i < 6; i++) {
-        addLinkedCoord(&loop, &sfVerts[i]);
-    }
+    createLinkedLoop(&loop, &sfVerts[0], 6);
 
     BBox bbox;
     bboxFromLinkedGeoLoop(&loop, &bbox);
@@ -199,14 +178,10 @@ TEST(bboxesFromGeoPolygonHole) {
 }
 
 TEST(bboxFromLinkedGeoLoop) {
-    const GeoCoord verts[] = {{0.8, 0.3}, {0.7, 0.6}, {1.1, 0.7}, {1.0, 0.2}};
+    GeoCoord verts[] = {{0.8, 0.3}, {0.7, 0.6}, {1.1, 0.7}, {1.0, 0.2}};
 
     LinkedGeoLoop loop;
-    initLinkedLoop(&loop);
-
-    for (int i = 0; i < 4; i++) {
-        addLinkedCoord(&loop, &verts[i]);
-    }
+    createLinkedLoop(&loop, &verts[0], 4);
 
     const BBox expected = {1.1, 0.7, 0.7, 0.2};
 
@@ -239,14 +214,9 @@ TEST(isClockwiseGeofence) {
 }
 
 TEST(isClockwiseLinkedGeoLoop) {
-    const GeoCoord verts[] = {{0.1, 0.1}, {0.2, 0.2}, {0.1, 0.2}};
-
+    GeoCoord verts[] = {{0.1, 0.1}, {0.2, 0.2}, {0.1, 0.2}};
     LinkedGeoLoop loop;
-    initLinkedLoop(&loop);
-
-    for (int i = 0; i < 3; i++) {
-        addLinkedCoord(&loop, &verts[i]);
-    }
+    createLinkedLoop(&loop, &verts[0], 3);
 
     t_assert(isClockwiseLinkedGeoLoop(&loop), "Got true for clockwise loop");
 
@@ -254,14 +224,9 @@ TEST(isClockwiseLinkedGeoLoop) {
 }
 
 TEST(isNotClockwiseLinkedGeoLoop) {
-    const GeoCoord verts[] = {{0, 0}, {0, 0.4}, {0.4, 0.4}, {0.4, 0}};
-
+    GeoCoord verts[] = {{0, 0}, {0, 0.4}, {0.4, 0.4}, {0.4, 0}};
     LinkedGeoLoop loop;
-    initLinkedLoop(&loop);
-
-    for (int i = 0; i < 4; i++) {
-        addLinkedCoord(&loop, &verts[i]);
-    }
+    createLinkedLoop(&loop, &verts[0], 4);
 
     t_assert(!isClockwiseLinkedGeoLoop(&loop),
              "Got false for counter-clockwise loop");
@@ -269,18 +234,23 @@ TEST(isNotClockwiseLinkedGeoLoop) {
     destroyLinkedGeoLoop(&loop);
 }
 
+TEST(isClockwiseGeofenceTransmeridian) {
+    GeoCoord verts[] = {{0.4, M_PI - 0.1},
+                        {0.4, -M_PI + 0.1},
+                        {-0.4, -M_PI + 0.1},
+                        {-0.4, M_PI - 0.1}};
+    Geofence geofence = {.numVerts = 4, .verts = verts};
+
+    t_assert(isClockwiseGeofence(&geofence), "Got true for clockwise geofence");
+}
+
 TEST(isClockwiseLinkedGeoLoopTransmeridian) {
-    const GeoCoord verts[] = {{0.4, M_PI - 0.1},
-                              {0.4, -M_PI + 0.1},
-                              {-0.4, -M_PI + 0.1},
-                              {-0.4, M_PI - 0.1}};
-
+    GeoCoord verts[] = {{0.4, M_PI - 0.1},
+                        {0.4, -M_PI + 0.1},
+                        {-0.4, -M_PI + 0.1},
+                        {-0.4, M_PI - 0.1}};
     LinkedGeoLoop loop;
-    initLinkedLoop(&loop);
-
-    for (int i = 0; i < 4; i++) {
-        addLinkedCoord(&loop, &verts[i]);
-    }
+    createLinkedLoop(&loop, &verts[0], 4);
 
     t_assert(isClockwiseLinkedGeoLoop(&loop),
              "Got true for clockwise transmeridian loop");
@@ -289,22 +259,298 @@ TEST(isClockwiseLinkedGeoLoopTransmeridian) {
 }
 
 TEST(isNotClockwiseLinkedGeoLoopTransmeridian) {
-    const GeoCoord verts[] = {{0.4, M_PI - 0.1},
-                              {-0.4, M_PI - 0.1},
-                              {-0.4, -M_PI + 0.1},
-                              {0.4, -M_PI + 0.1}};
-
+    GeoCoord verts[] = {{0.4, M_PI - 0.1},
+                        {-0.4, M_PI - 0.1},
+                        {-0.4, -M_PI + 0.1},
+                        {0.4, -M_PI + 0.1}};
     LinkedGeoLoop loop;
-    initLinkedLoop(&loop);
-
-    for (int i = 0; i < 4; i++) {
-        addLinkedCoord(&loop, &verts[i]);
-    }
+    createLinkedLoop(&loop, &verts[0], 4);
 
     t_assert(!isClockwiseLinkedGeoLoop(&loop),
              "Got false for counter-clockwise transmeridian loop");
 
     destroyLinkedGeoLoop(&loop);
+}
+
+TEST(normalizeMultiPolygonSingle) {
+    GeoCoord verts[] = {{0, 0}, {0, 1}, {1, 1}};
+
+    LinkedGeoLoop* outer = calloc(1, sizeof(*outer));
+    assert(outer != NULL);
+    createLinkedLoop(outer, &verts[0], 3);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, outer);
+
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_SUCCESS, "No error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 1, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 1, "Loop count correct");
+    t_assert(polygon.first == outer, "Got expected loop");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(normalizeMultiPolygonTwoOuterLoops) {
+    GeoCoord verts1[] = {{0, 0}, {0, 1}, {1, 1}};
+
+    LinkedGeoLoop* outer1 = calloc(1, sizeof(*outer1));
+    assert(outer1 != NULL);
+    createLinkedLoop(outer1, &verts1[0], 3);
+
+    GeoCoord verts2[] = {{2, 2}, {2, 3}, {3, 3}};
+
+    LinkedGeoLoop* outer2 = calloc(1, sizeof(*outer2));
+    assert(outer2 != NULL);
+    createLinkedLoop(outer2, &verts2[0], 3);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, outer1);
+    addLinkedLoop(&polygon, outer2);
+
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_SUCCESS, "No error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 2, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 1,
+             "Loop count on first polygon correct");
+    t_assert(countLinkedLoops(polygon.next) == 1,
+             "Loop count on second polygon correct");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(normalizeMultiPolygonOneHole) {
+    GeoCoord verts[] = {{0, 0}, {0, 3}, {3, 3}, {3, 0}};
+
+    LinkedGeoLoop* outer = calloc(1, sizeof(*outer));
+    assert(outer != NULL);
+    createLinkedLoop(outer, &verts[0], 4);
+
+    GeoCoord verts2[] = {{1, 1}, {2, 2}, {1, 2}};
+
+    LinkedGeoLoop* inner = calloc(1, sizeof(*inner));
+    assert(inner != NULL);
+    createLinkedLoop(inner, &verts2[0], 3);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, inner);
+    addLinkedLoop(&polygon, outer);
+
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_SUCCESS, "No error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 1, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 2,
+             "Loop count on first polygon correct");
+    t_assert(polygon.first == outer, "Got expected outer loop");
+    t_assert(polygon.first->next == inner, "Got expected inner loop");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(normalizeMultiPolygonTwoHoles) {
+    GeoCoord verts[] = {{0, 0}, {0, 0.4}, {0.4, 0.4}, {0.4, 0}};
+
+    LinkedGeoLoop* outer = calloc(1, sizeof(*outer));
+    assert(outer != NULL);
+    createLinkedLoop(outer, &verts[0], 4);
+
+    GeoCoord verts2[] = {{0.1, 0.1}, {0.2, 0.2}, {0.1, 0.2}};
+
+    LinkedGeoLoop* inner1 = calloc(1, sizeof(*inner1));
+    assert(inner1 != NULL);
+    createLinkedLoop(inner1, &verts2[0], 3);
+
+    GeoCoord verts3[] = {{0.2, 0.2}, {0.3, 0.3}, {0.2, 0.3}};
+
+    LinkedGeoLoop* inner2 = calloc(1, sizeof(*inner2));
+    assert(inner2 != NULL);
+    createLinkedLoop(inner2, &verts3[0], 3);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, inner2);
+    addLinkedLoop(&polygon, outer);
+    addLinkedLoop(&polygon, inner1);
+
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_SUCCESS, "No error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 1,
+             "Polygon count correct for 2 holes");
+    t_assert(polygon.first == outer, "Got expected outer loop");
+    t_assert(countLinkedLoops(&polygon) == 3,
+             "Loop count on first polygon correct");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(normalizeMultiPolygonTwoDonuts) {
+    GeoCoord verts[] = {{0, 0}, {0, 3}, {3, 3}, {3, 0}};
+    LinkedGeoLoop* outer = calloc(1, sizeof(*outer));
+    assert(outer != NULL);
+    createLinkedLoop(outer, &verts[0], 4);
+
+    GeoCoord verts2[] = {{1, 1}, {2, 2}, {1, 2}};
+    LinkedGeoLoop* inner = calloc(1, sizeof(*inner));
+    assert(inner != NULL);
+    createLinkedLoop(inner, &verts2[0], 3);
+
+    GeoCoord verts3[] = {{0, 0}, {0, -3}, {-3, -3}, {-3, 0}};
+    LinkedGeoLoop* outer2 = calloc(1, sizeof(*outer));
+    assert(outer2 != NULL);
+    createLinkedLoop(outer2, &verts3[0], 4);
+
+    GeoCoord verts4[] = {{-1, -1}, {-2, -2}, {-1, -2}};
+    LinkedGeoLoop* inner2 = calloc(1, sizeof(*inner));
+    assert(inner2 != NULL);
+    createLinkedLoop(inner2, &verts4[0], 3);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, inner2);
+    addLinkedLoop(&polygon, inner);
+    addLinkedLoop(&polygon, outer);
+    addLinkedLoop(&polygon, outer2);
+
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_SUCCESS, "No error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 2, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 2,
+             "Loop count on first polygon correct");
+    t_assert(countLinkedCoords(polygon.first) == 4, "Got expected outer loop");
+    t_assert(countLinkedCoords(polygon.first->next) == 3,
+             "Got expected inner loop");
+    t_assert(countLinkedLoops(polygon.next) == 2,
+             "Loop count on second polygon correct");
+    t_assert(countLinkedCoords(polygon.next->first) == 4,
+             "Got expected outer loop");
+    t_assert(countLinkedCoords(polygon.next->first->next) == 3,
+             "Got expected inner loop");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(normalizeMultiPolygonNestedDonuts) {
+    GeoCoord verts[] = {{0.2, 0.2}, {0.2, -0.2}, {-0.2, -0.2}, {-0.2, 0.2}};
+    LinkedGeoLoop* outer = calloc(1, sizeof(*outer));
+    assert(outer != NULL);
+    createLinkedLoop(outer, &verts[0], 4);
+
+    GeoCoord verts2[] = {{0.1, 0.1}, {-0.1, 0.1}, {-0.1, -0.1}, {0.1, -0.1}};
+    LinkedGeoLoop* inner = calloc(1, sizeof(*inner));
+    assert(inner != NULL);
+    createLinkedLoop(inner, &verts2[0], 4);
+
+    GeoCoord verts3[] = {{0.6, 0.6}, {0.6, -0.6}, {-0.6, -0.6}, {-0.6, 0.6}};
+    LinkedGeoLoop* outerBig = calloc(1, sizeof(*outerBig));
+    assert(outerBig != NULL);
+    createLinkedLoop(outerBig, &verts3[0], 4);
+
+    GeoCoord verts4[] = {{0.5, 0.5}, {-0.5, 0.5}, {-0.5, -0.5}, {0.5, -0.5}};
+    LinkedGeoLoop* innerBig = calloc(1, sizeof(*innerBig));
+    assert(innerBig != NULL);
+    createLinkedLoop(innerBig, &verts4[0], 4);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, inner);
+    addLinkedLoop(&polygon, outerBig);
+    addLinkedLoop(&polygon, innerBig);
+    addLinkedLoop(&polygon, outer);
+
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_SUCCESS, "No error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 2, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 2,
+             "Loop count on first polygon correct");
+    t_assert(polygon.first == outerBig, "Got expected outer loop");
+    t_assert(polygon.first->next == innerBig, "Got expected inner loop");
+    t_assert(countLinkedLoops(polygon.next) == 2,
+             "Loop count on second polygon correct");
+    t_assert(polygon.next->first == outer, "Got expected outer loop");
+    t_assert(polygon.next->first->next == inner, "Got expected inner loop");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(normalizeMultiPolygonNoOuterLoops) {
+    GeoCoord verts1[] = {{0, 0}, {1, 1}, {0, 1}};
+
+    LinkedGeoLoop* outer1 = calloc(1, sizeof(*outer1));
+    assert(outer1 != NULL);
+    createLinkedLoop(outer1, &verts1[0], 3);
+
+    GeoCoord verts2[] = {{2, 2}, {3, 3}, {2, 3}};
+
+    LinkedGeoLoop* outer2 = calloc(1, sizeof(*outer2));
+    assert(outer2 != NULL);
+    createLinkedLoop(outer2, &verts2[0], 3);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, outer1);
+    addLinkedLoop(&polygon, outer2);
+
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_ERR_UNASSIGNED_HOLES,
+             "Expected error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 1, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 0,
+             "Loop count as expected with invalid input");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
+}
+
+TEST(normalizeMultiPolygonAlreadyNormalized) {
+    GeoCoord verts1[] = {{0, 0}, {0, 1}, {1, 1}};
+
+    LinkedGeoLoop* outer1 = calloc(1, sizeof(*outer1));
+    assert(outer1 != NULL);
+    createLinkedLoop(outer1, &verts1[0], 3);
+
+    GeoCoord verts2[] = {{2, 2}, {2, 3}, {3, 3}};
+
+    LinkedGeoLoop* outer2 = calloc(1, sizeof(*outer2));
+    assert(outer2 != NULL);
+    createLinkedLoop(outer2, &verts2[0], 3);
+
+    LinkedGeoPolygon polygon;
+    initLinkedPolygon(&polygon);
+    addLinkedLoop(&polygon, outer1);
+    LinkedGeoPolygon* next = addNewLinkedPolygon(&polygon);
+    addLinkedLoop(next, outer2);
+
+    // Should be a no-op
+    int result = normalizeMultiPolygon(&polygon);
+
+    t_assert(result == NORMALIZATION_ERR_MULTIPLE_POLYGONS,
+             "Expected error code returned");
+
+    t_assert(countLinkedPolygons(&polygon) == 2, "Polygon count correct");
+    t_assert(countLinkedLoops(&polygon) == 1,
+             "Loop count on first polygon correct");
+    t_assert(polygon.first == outer1, "Got expected outer loop");
+    t_assert(countLinkedLoops(polygon.next) == 1,
+             "Loop count on second polygon correct");
+    t_assert(polygon.next->first == outer2, "Got expected outer loop");
+
+    H3_EXPORT(destroyLinkedPolygon)(&polygon);
 }
 
 END_TESTS();