+++ src/h3lib/include/linkedGeo.h
@@ -26,11 +26,13 @@
 
 void initLinkedPolygon(LinkedGeoPolygon* polygon);
 void initLinkedLoop(LinkedGeoLoop* loop);
-LinkedGeoPolygon* addLinkedPolygon(LinkedGeoPolygon* polygon);
-LinkedGeoLoop* addLinkedLoop(LinkedGeoPolygon* polygon);
+LinkedGeoPolygon* addNewLinkedPolygon(LinkedGeoPolygon* polygon);
+LinkedGeoLoop* addNewLinkedLoop(LinkedGeoPolygon* polygon);
+LinkedGeoLoop* addLinkedLoop(LinkedGeoPolygon* polygon, LinkedGeoLoop* loop);
 LinkedGeoCoord* addLinkedCoord(LinkedGeoLoop* loop, const GeoCoord* vertex);
 int countLinkedPolygons(LinkedGeoPolygon* polygon);
 int countLinkedLoops(LinkedGeoPolygon* polygon);
 int countLinkedCoords(LinkedGeoLoop* loop);
+void destroyLinkedGeoLoop(LinkedGeoLoop* loop);
 
 #endif